feat: chat empty state extraction and boundary cleanup
